記事タイトルは別コンポーネントで表示させるように変更
