+++ src/styles/blog.module.css
@@ -1,3 +1,10 @@
+.page_title {
+  font-size: 32px;
+  font-weight: normal;
+  margin: 60px auto;
+  text-align: center;
+}
+
 .blog {
   max-width: 820px;
   margin: auto;
