+++ src/styles/blogTitle.module.css
@@ -1,10 +1,3 @@
-.page_title {
-  font-size: 32px;
-  font-weight: normal;
-  margin: 60px auto;
-  text-align: center;
-}
-
 .article_list {
   list-style: none;
   margin: 20px auto;
